Tidy up and catch already installed

+++ build/mongo/sharded/mongo_cluster.sh
@@ -24,17 +24,21 @@ chown mongodb:mongodb /etc/mongo-shards-down
 chmod a+x /etc/mongo-shards-up
 
 
+if [ ! -f /opt/mongo_shard_perms.done ]
+then
 echo " - set up mongo permissions"
+#Add oustudent user to mongodb group
 usermod -a -G mongodb oustudent
-#Grant permissions on certain commands
+#Grant permissions on certain commands to mongodb group
 echo '''
-%mongodb ALL=NOPASSWD: /usr/bin/killall mongod
-%mongodb ALL=NOPASSWD: /usr/bin/killall mongos
-%mongodb ALL=NOPASSWD: /bin/systemctl restart mongodb
-%mongodb ALL=NOPASSWD: /etc/mongo-shards-down
-%mongodb ALL=NOPASSWD: /etc/mongo-shards-up
+%mongodb    ALL=NOPASSWD:   /usr/bin/killall mongod
+%mongodb    ALL=NOPASSWD:   /usr/bin/killall mongos
+%mongodb    ALL=NOPASSWD:   /bin/systemctl restart mongodb
+%mongodb    ALL=NOPASSWD:   /etc/mongo-shards-down
+%mongodb    ALL=NOPASSWD:   /etc/mongo-shards-up
 '''  >> /etc/sudoers
-
+touch /opt/mongo_shard_perms.done
+fi
 #--------------
 
 echo "killing mongod and mongos"
@@ -70,49 +74,51 @@ mkdir -p /data/config/config-a
 chmod -R a+rwx  /data/config/config-a
 
 ##
+if [ ! -f /opt/mongo_shard_restore.done ]; then
+    #Fire up the servers
+    /etc/mongo-shards-up
 
-#Fire up the servers
-/etc/mongo-shards-up
-
-#add the shards
-echo "Connecting to mongos and enabling sharding"
-# add shards and enable sharding on the test db
-mongo < ${THISDIR}/createshards.js
+    #add the shards
+    echo "Connecting to mongos and enabling sharding"
+    # add shards and enable sharding on the test db
+    mongo < ${THISDIR}/createshards.js
 
 
-#--------- SEED THE DATABASE WITH SOME DATA
+    #--------- SEED THE DATABASE WITH SOME DATA
 
-##Copy across and unpack the data files into the VM
+    ##Copy across and unpack the data files into the VM
 
-#Example of creating tar.bz2 file: 
-# tar -cvjSf filename.tar.bz2 dirname
+    #Example of creating tar.bz2 file: 
+    # tar -cvjSf filename.tar.bz2 dirname
 
-echo "Copy some data across (if required)..."
-if [ ! -f /root/accidents-0912.tar.bz2 ]; then
-	cp $THISDIR/data/accidents-0912.tar.bz2 /root/accidents-0912.tar.bz2
+    echo "Copy some data across (if required)..."
+    if [ ! -f /root/accidents-0912.tar.bz2 ]; then
+        cp $THISDIR/data/accidents-0912.tar.bz2 /root/accidents-0912.tar.bz2
+    fi
+    echo "...data copied"
+
+    echo "Unpack some data (if required)..."
+    if [ ! -f /tmpdatafiles/dump-0912 ]; then
+        mkdir -p /tmpdatafiles
+        tar xvjf /root/accidents-0912.tar.bz2 -C /tmpdatafiles
+    fi
+    echo "...unpacking done"
+
+    echo "Try a restore..."
+    #Should we do this with --drop ?
+    #https://docs.mongodb.com/manual/reference/program/mongorestore/#cmdoption--drop
+    mongorestore --numInsertionWorkersPerCollection 2 --port=$MONGOD_PORT /tmpdatafiles/dump-0912 && rm -rf /tmpdatafiles
+    echo "...restore done"
+
+    
+
+    echo "...set up mongo cluster done"
+
+    #fi
+
+    /etc/mongo-shards-down
+
+    touch /opt/mongo_shard_restore.done
 fi
-echo "...data copied"
-
-echo "Unpack some data (if required)..."
-if [ ! -f /tmpdatafiles/dump-0912 ]; then
-	mkdir -p /tmpdatafiles
-	tar xvjf /root/accidents-0912.tar.bz2 -C /tmpdatafiles
-fi
-echo "...unpacking done"
-
-echo "Try a restore..."
-#Should we do this with --drop ?
-#https://docs.mongodb.com/manual/reference/program/mongorestore/#cmdoption--drop
-mongorestore --port=$MONGOD_PORT /tmpdatafiles/dump-0912
-echo "...restore done"
-
-##Tidy up
-rm -rf /tmpdatafiles
-
-echo "...set up mongo cluster"
-
-#fi
-
-/etc/mongo-shards-down
 
 touch /opt/mongo_cluster.done
